Add interface and default implementation for AWS Kinesis client factory
Working within the log4net conventions does not allow for injecting
appender dependencies (e.g., the AWS Kinesis client) via the
constructor. Instead we can inject the client via a property. However,
since the convention of the AWS SDK is to use a static method on the
AWSClientFactory class to create a client instance instead of directly
instantiating the client instance, we will create a property on the
appender for the client factory instead of the client.